Make periodo app into a package with separate submodules.
