+++ .gitignore
@@ -9,7 +9,7 @@ include/*
 lib/*
 local
 local/*
-static/html
 
 db.sqlite
-secrets.py
+periodo/secrets.py
+periodo/static/html
