+++ periodo/static/vocab.ttl
