+++ test/test_patch.py
@@ -1,7 +1,5 @@
 import os
 import json
-import periodo
-import identifier
 import tempfile
 import unittest
 import http.client
@@ -10,6 +8,7 @@ from rdflib.namespace import Namespace
 from urllib.parse import urlparse
 from flask.ext.principal import ActionNeed
 from .filepath import filepath
+from periodo import app, database, identifier, commands, auth
 
 PERIODO = Namespace('http://n2t.net/ark:/99152/')
 PROV = Namespace('http://www.w3.org/ns/prov#')
@@ -18,45 +17,47 @@ PROV = Namespace('http://www.w3.org/ns/prov#')
 class TestPatchMethods(unittest.TestCase):
 
     def setUp(self):
-        self.db_fd, periodo.app.config['DATABASE'] = tempfile.mkstemp()
-        periodo.app.config['TESTING'] = True
-        self.app = periodo.app.test_client()
-        periodo.init_db()
-        periodo.load_data(filepath('test-data.json'))
-        self.user_identity = periodo.add_user_or_update_credentials({
-            'name': 'Regular Gal',
-            'access_token': '5005eb18-be6b-4ac0-b084-0443289b3378',
-            'expires_in': 631138518,
-            'orcid': '1234-5678-9101-112X',
-        })
-        self.admin_identity = periodo.add_user_or_update_credentials({
-            'name': 'Super Admin',
-            'access_token': 'f7c64584-0750-4cb6-8c81-2932f5daabb8',
-            'expires_in': 3600,
-            'orcid': '1211-1098-7654-321X',
-        }, (ActionNeed('accept-patch'),))
+        self.db_fd, app.config['DATABASE'] = tempfile.mkstemp()
+        app.config['TESTING'] = True
+        self.client = app.test_client()
         with open(filepath('test-patch-replace-values-1.json')) as f:
             self.patch = f.read()
+        commands.init_db()
+        commands.load_data(filepath('test-data.json'))
+        with app.app_context():
+            self.user_identity = auth.add_user_or_update_credentials({
+                'name': 'Regular Gal',
+                'access_token': '5005eb18-be6b-4ac0-b084-0443289b3378',
+                'expires_in': 631138518,
+                'orcid': '1234-5678-9101-112X',
+            })
+            self.admin_identity = auth.add_user_or_update_credentials({
+                'name': 'Super Admin',
+                'access_token': 'f7c64584-0750-4cb6-8c81-2932f5daabb8',
+                'expires_in': 3600,
+                'orcid': '1211-1098-7654-321X',
+            }, (ActionNeed('accept-patch'),))
+            database.commit()
 
     def tearDown(self):
         os.close(self.db_fd)
-        os.unlink(periodo.app.config['DATABASE'])
+        os.unlink(app.config['DATABASE'])
 
     def test_initial_patch(self):
-        with self.app as client:
+        with self.client as client:
             client.patch(
                 '/d/',
                 data=self.patch,
                 content_type='application/json',
                 headers={'Authorization': 'Bearer '
                          + 'NTAwNWViMTgtYmU2Yi00YWMwLWIwODQtMDQ0MzI4OWIzMzc4'})
-            affected_entities = periodo.query_db(
+            affected_entities = database.query_db(
                 'SELECT affected_entities FROM patch_request WHERE id = 1',
                 one=True)['affected_entities']
             self.assertEqual(affected_entities, '["p0trgkv", "p0trgkvwbjd"]')
 
     def test_submit_patch(self):
-        with self.app as client:
+        with self.client as client:
             res = client.patch(
                 '/d/',
                 data=self.patch,
@@ -65,13 +66,13 @@ class TestPatchMethods(unittest.TestCase):
                          + 'NTAwNWViMTgtYmU2Yi00YWMwLWIwODQtMDQ0MzI4OWIzMzc4'})
             self.assertEqual(res.status_code, http.client.ACCEPTED)
             patch_id = int(res.headers['Location'].split('/')[-2])
-            affected_entities = periodo.query_db(
+            affected_entities = database.query_db(
                 'SELECT affected_entities FROM patch_request WHERE id = ?',
                 (patch_id,), one=True)['affected_entities']
             self.assertEqual(affected_entities, '["p0trgkv", "p0trgkvwbjd"]')
 
     def test_update_patch(self):
-        res = self.app.patch(
+        res = self.client.patch(
             '/d/',
             data=self.patch,
             content_type='application/json',
@@ -80,26 +81,26 @@ class TestPatchMethods(unittest.TestCase):
         patch_url = urlparse(res.headers['Location']).path
         jsonpatch_url = patch_url + 'patch.jsonpatch'
 
-        res = self.app.get(jsonpatch_url)
+        res = self.client.get(jsonpatch_url)
         self.assertEqual(json.loads(self.patch),
                          json.loads(res.get_data(as_text=True)))
         with open(filepath('test-patch-replace-values-2.json')) as f:
             self.patch2 = f.read()
-        res = self.app.put(
+        res = self.client.put(
             jsonpatch_url,
             data=self.patch2,
             content_type='application/json',
             headers={'Authorization': 'Bearer '
                          + 'NTAwNWViMTgtYmU2Yi00YWMwLWIwODQtMDQ0MzI4OWIzMzc4'})
         self.assertEqual(res.status_code, http.client.OK)
-        res = self.app.get(jsonpatch_url)
+        res = self.client.get(jsonpatch_url)
         self.assertEqual(json.loads(self.patch2),
                          json.loads(res.get_data(as_text=True)))
 
     def test_merge_patch(self):
         with open(filepath('test-patch-adds-items.json')) as f:
             patch = f.read()
-        with self.app as client:
+        with self.client as client:
             res = client.patch(
                 '/d/',
                 data=patch,
@@ -107,7 +108,7 @@ class TestPatchMethods(unittest.TestCase):
                 headers={'Authorization': 'Bearer '
                          + 'NTAwNWViMTgtYmU2Yi00YWMwLWIwODQtMDQ0MzI4OWIzMzc4'})
             patch_id = int(res.headers['Location'].split('/')[-2])
-            affected_entities = periodo.query_db(
+            affected_entities = database.query_db(
                 'SELECT affected_entities FROM patch_request WHERE id = ?',
                 (patch_id,), one=True)['affected_entities']
             self.assertEqual(affected_entities, '["p0trgkv"]')
@@ -117,12 +118,12 @@ class TestPatchMethods(unittest.TestCase):
                 headers={'Authorization': 'Bearer '
                          + 'ZjdjNjQ1ODQtMDc1MC00Y2I2LThjODEtMjkzMmY1ZGFhYmI4'})
             self.assertEqual(res.status_code, http.client.NO_CONTENT)
-            row = periodo.query_db(
+            row = database.query_db(
                 'SELECT applied_to, resulted_in FROM patch_request WHERE id=?',
                 (patch_id,), one=True)
             self.assertEqual(1, row['applied_to'])
             self.assertEqual(2, row['resulted_in'])
-            affected_entities = json.loads(periodo.query_db(
+            affected_entities = json.loads(database.query_db(
                 'SELECT affected_entities FROM patch_request WHERE id = ?',
                 (patch_id,), one=True)['affected_entities'])
             self.assertEqual(4, len(affected_entities))
@@ -134,7 +135,7 @@ class TestPatchMethods(unittest.TestCase):
             patch1 = f.read()
         with open(filepath('test-patch-add-definition.json')) as f:
             patch2 = f.read()
-        with self.app as client:
+        with self.client as client:
             res = client.patch(
                 '/d/',
                 data=patch1,
@@ -225,7 +226,8 @@ class TestPatchMethods(unittest.TestCase):
                     '/' + res.headers['Location'].split('/')[-1],
                     '{}.json?version=3'.format(path))
                 res = client.get('{}.json?version=3'.format(path))
-                self.assertEqual(res.status_code, http.client.MOVED_PERMANENTLY)
+                self.assertEqual(
+                    res.status_code, http.client.MOVED_PERMANENTLY)
                 self.assertEqual(
                     '/' + res.headers['Location'].split('/')[-1],
                     '{}.json?version=2'.format(path))
